07-presentation/demo/01/requirements.txt
openai
python-dotenv
requests
streamlit
gradio
